refactor: better handle creation of test env
